+++ test/test_file.py
@@ -9,10 +9,11 @@ class FileConversionMixin(FormatMixin):
         dest = self.DEST
         source_file = self.test_data_path('test_file.%s' % source)
         expected_file = self.test_data_path('test_file.%s' % dest)
-        result_file = self.test_data_path('_%s'%source, 'test_file.%s' % dest)
+        result_file = self.test_data_path('_%s' % source, 'test_file.%s' % dest)
         cli_converter([source_file, result_file])
         self.assert_result_file(expected_file, result_file,
-                                "Failure converting %s into %s"%(source, dest))
+                                "Failure converting %s into %s" % (
+                                    source, dest))
         self.remove_file(result_file)
 
     def test_html(self):
@@ -51,31 +52,39 @@ class FileConversionTestPsql(BaseTest, FileConversionMixin):
     DEST = 'psql'
 
 
-class ParseArgsColumnsTest(BaseTest):
+class ParseArgsTest(BaseTest):
     def validate_test_condition(self, name, source, dest, *columns):
         source_file = self.test_data_path('test_file.%s' % source)
         expected_file = self.test_data_path('expected', '%s.%s' % (name, dest))
-        result_file = self.test_data_path('_parse_args', '%s.%s'%(name, dest))
-        cli_converter([source_file, result_file, '--columns'] + list(columns))
+        result_file = self.test_data_path('_parse_args', '%s.%s' % (name, dest))
+        cli_converter([source_file, result_file] + list(columns))
         self.assert_result_file(expected_file, result_file,
-                                "Failure converting %s into %s"%(source, dest))
+                                "Failure converting %s into %s" % (
+                                    source, dest))
         self.remove_file(result_file)
 
     def test_reorder_columns(self):
-        self.validate_test_condition('test_reorder_columns',
-                                     'txt', 'grid',
-                                     'Target', 'Action', 'Role', 'Player', 'TU')
+        self.validate_test_condition('test_reorder_columns', 'txt', 'grid',
+                                     '--columns', 'Target', 'Action', 'Role',
+                                     'Player', 'TU')
 
     def test_remove_columns(self):
-        self.validate_test_condition('test_recorder_columns',
-                                     'json', 'psql',
-                                     'TU', 'Player', 'Action')
+        self.validate_test_condition('test_remove_columns', 'json', 'psql',
+                                     '--columns', 'TU', 'Player', 'Action')
 
     def test_add_columns(self):
-        self.validate_test_condition('test_recorder_columns',
-                                     'csv', 'md',
-                                     'TU', 'Addition1', 'Player', 'Action',
-                                     'Role', 'Player', 'TU', 'Addition2')
+        self.validate_test_condition('test_add_columns', 'csv', 'md',
+                                     '--columns', 'TU', 'Addition1', 'Player',
+                                     'Action', 'Role', 'Player', 'TU',
+                                     'Addition2')
+
+    def test_reorder_rows(self):
+        self.validate_test_condition('test_reorder_rows', 'csv', 'md',
+                                     '--order-by', 'TU', 'Role')
+
+    def test_reorder_rows_reverse(self):
+        self.validate_test_condition('test_reorder_rows_reverse', 'csv', 'md',
+                                     '--order-by', '~Player', 'Target')
 
 
 if __name__ == '__main__':