+++ CHANGELOG.rst
@@ -1,8 +1,11 @@
 Changes in 6.2.0
 ================
 
-* Replaced manual parsing args with ArgumentParser and added a ``--column``
-  option.
+* Replaced manual parsing args with ArgumentParser.
+
+* Added ``--column`` command line argument to specify the new columns.
+
+* Added ``--order-by`` command line argument to specify how to sort the rows.
 
 
 Changes in 6.1.0